Fix to token creation And unit test surrounding it
